Add shortcode implementations
Implementations for the `modal` and `modal-launch` shortcodes added. The
main basis of these implementations is the jquery-modal library.

A custom version of the jquery-modal library is used, which changes the
namespacing of the JavaScript objects and classes so that they are less
likely to conflict with frontend frameworks typically used by Grav
themes. The source of this custom version is the tag 'v0.9.1-jqmodal'
from this fork of the main repo: https://github.com/marshmn/jquery-modal